+++ modulo6/revisao-fullstack/frontend/src/components/Loading/Loading.jsx
@@ -1,0 +1,8 @@
+import React from "react"
+import {LoadingSection} from './style'
+
+export function Loading () {
+    return (
+        <LoadingSection></LoadingSection>
+    )
+}
