+++ modulo6/revisao-fullstack/frontend/src/components/ProductsInCart/style.jsx
@@ -1,0 +1,20 @@
+import styled from "styled-components"
+
+
+export const Container = styled.section `
+display: flex;
+flex-direction: column;
+align-items: center;
+gap: 1.5vh;
+
+h3 {
+    font-size: 24px;
+    margin-top: 3vh;
+}
+
+#totalPrice {
+    font-size: 26px;
+    color: red;
+    margin-top: 1vh;
+}
+`
